requirements.txt
paho-mqtt
google-cloud-vision
